+++ testauth/settings/local.py
@@ -61,6 +61,7 @@ if os.environ.get("USE_MYSQL", True) is True:
 
 # Add any additional apps to this list.
 INSTALLED_APPS += [
+    "allianceauth.services.modules.discord",
     PACKAGE,
     "aadiscordbot",
 ]